The RentManager has added a new business rule and updated the controls.

+++ src/main/java/com/nexgencarrental/nexGenCarRental/entities/Car.java
@@ -20,7 +20,7 @@ public class Car {
     private int id;
 
     @Column(name ="kilometer")
-    private int kilometer;
+    private double kilometer;
 
     @Column(name = "year")
     private int year;
