+++ src/main/java/com/nexgencarrental/nexGenCarRental/services/dtos/responses/car/GetCarListResponse.java
@@ -18,7 +18,7 @@ import java.util.List;
 @NoArgsConstructor
 public class GetCarListResponse {
     private String plate;
-    private int kilometer;
+    private double kilometer;
     private double dailyPrice;
     private int year;
     private String modelName;
